% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nbs_comparison.R
\name{add_crf_ids}
\alias{add_crf_ids}
\title{Add ID Columns to CRF Data}
\usage{
add_crf_ids(data = filter_crf())
}
\arguments{
\item{data}{Data from \code{filter_crf()}}
}
\value{
\code{data} with added columns, all with suffix \verb{_id_tmp_}
}
\description{
Add ID Columns to CRF Data
}
\keyword{internal}
