% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nbs_comparison.R
\name{add_nbs_ids}
\alias{add_nbs_ids}
\title{Add ID Columns to NBS Data}
\usage{
add_nbs_ids(data = load_positive())
}
\arguments{
\item{data}{Data from \code{filter_crf()}}
}
\value{
\code{data} with added columns, all with suffix \verb{_id_tmp_}
}
\description{
Add ID Columns to NBS Data
}
