% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crf_wb.R
\name{create_crf_wb}
\alias{create_crf_wb}
\title{Create Excel Workbook with Categorized Sheets of CRF Data}
\usage{
create_crf_wb(data = prep_crf())
}
\arguments{
\item{data}{CRF data prepared by \code{\link[covidcrf:prep_crf]{prep_crf()}}}
}
\value{
An openxlsx WorkBook object
}
\description{
\code{create_crf_wb()} creates an Excel workbook object using the
\href{https://ycphs.github.io/openxlsx/}{openxlsx} package. The data are
categorized into 5 sheets:
\describe{NBS Cases (NBS - No Recent Test)}{Records in NBS without a recent
test. New cases.}
\describe{NBS - Recent Test}{Records in NBS, but with another recent test.
Not new cases, since they've already tested positive recently.}
\describe{NBS - Test Status Unknown}{Records in NBS, but without sufficient
information to determine whether they've had a prior positive test recently.}
\describe{Not NBS}{Records not (yet) in NBS}
\describe{Not Determined}{Records without sufficient identifying information
to determine whether they're in NBS.}
}
