% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_crf.R
\name{distinct_crf}
\alias{distinct_crf}
\title{Deduplicate Cases and Coalesce Information Among Duplicates}
\usage{
distinct_crf(data = filter_crf())
}
\arguments{
\item{data}{A \code{tibble} with CRF data, usually output from
\code{\link[covidcrf:filter_crf]{filter_crf()}}}
}
\value{
The de-duplicated and coalesced input, sorted by \code{record_id}
}
\description{
\code{distinct_crf()} de-duplicates cases in the CRF data using first name,
last name, date of birth, and specimen collection date. Where possible,
missing data in one duplicate record is filled in from data in others.
}
