% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_crf.R
\name{download_crf}
\alias{download_crf}
\title{Download All Records from CRF REDcap Project}
\usage{
download_crf(api_token = Sys.getenv("redcap_CRF_token"))
}
\arguments{
\item{api_token}{API token/key for the CRF REDcap project}
}
\value{
A \code{tibble} in \code{character} format
}
\description{
\code{download_crf()} downloads all records from the CRF REDcap project and cleans
variables names with \code{\link[janitor:clean_names]{clean_names()}}. The
download will retry up to 12 times before failing. All variables are returned
as \code{character}.
}
