% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_crf.R
\name{filter_crf}
\alias{filter_crf}
\title{Filter to Cases of Interest}
\usage{
filter_crf(data = download_crf(), days = 5L, na_keep = TRUE)
}
\arguments{
\item{data}{A data frame with data from the CRF REDcap project, as output by
\code{\link[covidcrf:download_crf]{download_crf()}}}

\item{days}{The maxmimum allowed difference between last day worked
(\code{last_worked}) and test date (\code{specimendate})}

\item{na_keep}{Should missing data be kept or excluded?}
}
\value{
The input \code{data} filtered according to the description above
}
\description{
\code{filter_crf()} filters out cases which have previously answered the survey
or were not at work within \code{days} days of their onset/test date.
}
