% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_crf_archive.R
\name{manage_crf_archive}
\alias{manage_crf_archive}
\title{Manage CRF Data Archive}
\usage{
manage_crf_archive(dir = "V:/EPI DATA ANALYTICS TEAM/CRF/CRF_NBS", min = 7L)
}
\arguments{
\item{dir}{The CRF data directory}

\item{min}{The minimum number of files and days to keep}
}
\value{
The deleted paths (invisibly)
}
\description{
\code{manage_crf_archive()} deletes old version of the CRF data. By default,
it will retain at least the 7 most recent versions, and will retain all
versions created in the last 7 days.
}
