% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_crf.R
\name{map_crf_cols}
\alias{map_crf_cols}
\title{Map Raw Variable Names to Labels in CRF Data}
\usage{
map_crf_cols(cols, template = download_crf_template())
}
\arguments{
\item{cols}{Character. Raw column names in CRF data.}

\item{template}{The template for the CRF REDcap project}
}
\value{
\code{character} vector with raw column names replaced by column labels
}
\description{
Map Raw Variable Names to Labels in CRF Data
}
\keyword{internal}
