% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_crf_archive.R
\name{old_crf_files}
\alias{old_crf_files}
\title{Identify Old CRF Files}
\usage{
old_crf_files(
 
    dir = paste0("V:/OPERATIONS Surveillance Coordinator/Clusters/Case Report Forms/CRF_NBS"),
  regex = ".*/CRF_NBS[^/]*[.]xlsx$"
)
}
\arguments{
\item{dir}{Directory to search}

\item{regex}{Regular expression matching CRF files}
}
\value{
An \code{fs_path} vector
}
\description{
\code{old_crf_files()} identifies all CRF files in a directory except the most
recent
}
\keyword{internal}
