% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_crf.R
\name{prep_crf}
\alias{prep_crf}
\title{Prepare CRF Data for Export}
\usage{
prep_crf(data = add_recent_test(), template = download_crf_template())
}
\arguments{
\item{data}{Data ready for export preparation, as produced by
\code{\link[covidcrf:add_recent_test]{add_recent_test()}}}

\item{template}{The template for the CRF REDcap project}
}
\value{
The input \code{data} with export columns selected, renamed, and formatted
}
\description{
\code{prep_crf()} select relevant columns, renames those columns using labels from
the REDcap CRF project, and formats the contents of CRF data to prepare for
export to Excel. Non-export columns are dropped.
}
