% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_crf.R
\name{rename_crf}
\alias{rename_crf}
\title{Rename Export Columns in CRF Data}
\usage{
rename_crf(data = transmute_crf(), template = download_crf_template())
}
\arguments{
\item{data}{CRF data containing only export columns, as produced by
\code{\link[covidprod:transmute_crf]{transmute_crf()}}}

\item{template}{The template for the CRF REDcap project}
}
\value{
Input data with REDcap columns renamed
}
\description{
Rename Export Columns in CRF Data
}
\keyword{internal}
