% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_crf_wb.R
\name{save_crf_wb}
\alias{save_crf_wb}
\title{Save Workbook with CRF Data to Disk}
\usage{
save_crf_wb(
  wb = create_crf_wb(),
  dir = "V:/EPI DATA ANALYTICS TEAM/CRF/CRF_NBS",
  force = TRUE
)
}
\arguments{
\item{wb}{An openxlsx \code{Workbook} object, as output by \code{create_crf_wb()}}

\item{dir}{The directory to save the file}

\item{force}{Should an existing file at this location be overwritten? Unlike
most save functions in the covid* package universe, the default here is
\code{TRUE}.}
}
\value{
The path to the saved file (invisibly)
}
\description{
\code{save_crf_wb()} saves the output of
\code{\link[covidcrf:create_crf_wb]{create_crf_wb()}} to the specified
directory under the file name \code{crf_nbs_overlap_YYYY-MM-DD.xlsx}, where
\code{YYYY-MM-DD} is the date at the time of saving.
}
