% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_crf.R
\name{transmute_crf}
\alias{transmute_crf}
\title{Select and Mutate CRF Columns for Export}
\usage{
transmute_crf(data = add_recent_test())
}
\arguments{
\item{data}{Data ready for export preparation, as produced by
\code{\link[covidcrf:add_recent_test]{add_recent_test()}}}
}
\value{
The input \code{data} with export columns selected and renamed; other
columns are dropped
}
\description{
Select and Mutate CRF Columns for Export
}
\keyword{internal}
