% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magrittr.R
\name{utils-magrittr}
\alias{utils-magrittr}
\title{Magrittr Pipe Utilities}
\description{
The magrittr pipe is a common programming tool in \strong{R} that often makes
code much easier to read. In addition to the primary pipe operator, magrittr
offers several additional pipes, as well as functions to help perform common
infix-operations in a pipe-friendly manner.
}
\details{
These functions are currently bulk imported to allow easy reading of code
with the pipe operator. This decision may be re-thought at a later date.
}
\keyword{internal}
