% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nbs_comparison.R
\name{add_recent_test}
\alias{add_recent_test}
\title{Determine Whether Individuals Have Another Test Within \code{days}}
\usage{
add_recent_test(crf = crf_in_nbs(), nbs = load_positive(), days = 90L)
}
\arguments{
\item{crf}{Case Report Form data, usually output from
\code{\link[covidcrf:crf_in_nbs]{crf_in_nbs()}}}

\item{nbs}{Positive NBS data from
\code{\link[covidcrf:load_positive]{load_positive()}}}

\item{days}{Range of days for matching test dates}
}
\value{
The input \code{crf} data with a \code{logical} column \code{recent_test} added
}
\description{
\code{add_recent_test()} adds a column indicating whether a person has previously
tested positive within \code{days} of the given test date. The resulting
\code{recent_test} column is:
\itemize{
\item{\code{TRUE} if any previous tests are found}
\item{\code{FALSE} if no previous tests are found with complete data}
\item{\code{NA} if no previous tests are found with incomplete (missing) data}
}
}
